% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Simulate binary outcomes from a logistic model with center-level random
intercepts and year-within-center random intercepts}
\usage{
create_data(
  num_years = 5,
  max_num_centers = 400,
  alpha = -1.75,
  gamma_var = 0.5^2,
  omega_var = 0.15^2,
  omega_cor = 0.9,
  beta = 1,
  mean_runs_per_center = NULL
)
}
\arguments{
\item{num_years}{Number of years to simulate (integer, scalar)}

\item{max_num_centers}{Maximum possible number of centers to simulate. Since
runs, i.e. observations, within a center within a year are generated from a
Poisson, a given center may have no run in a given year (integer, scalar)}

\item{alpha}{Global intercept. (double, scalar)}

\item{gamma_var}{Variance of the center-level random intercepts (double,
scalar, positive)}

\item{omega_var}{Variance of the year-within-center random intercepts
(double, scalar, positive)}

\item{omega_cor}{AR(1) correlation of the year-within-center random
intercepts (double, scalar, in (0, 1))}

\item{beta}{Regression coefficient for the run-specific risk factor, which is
N(0,1) (double, scalar)}

\item{mean_runs_per_center}{Optional vector as long as \code{max_number_centers}
corresponding to the mean of the Poisson distribution. If provided, the
number of runs from each center from each year will be independently
generated from a Poisson distribution with this mean. If not provided, this
is created from a shifted half-t distribution to mimic right-skewness, i.e.
many centers with small numbers of runs. (double, vector, positive)}
}
\value{
A data.frame
}
\description{
Simulate binary outcomes from a logistic model with center-level random
intercepts and year-within-center random intercepts
}
\examples{
create_data()
}
